Add files via upload
This chatbot project is  developed using Python and the Tkinter library for the graphical user interface. This interactive chatbot is designed to engage users in conversations related to Virat Kohli, the renowned cricketer. The bot utilizes natural language processing techniques from the NLTK library, including lemmatization and tokenization, to understand and respond to user queries effectively. The chatbot employs a TF-IDF (Term Frequency-Inverse Document Frequency) vectorizer and cosine similarity to generate contextually relevant responses based on a dataset. Users can interact with the chatbot by typing messages in the provided entry field, and the bot responds with informative answers , creating an engaging and user-friendly experience.